LPS-174018 Move registry and unregistry methods to ObjectLayoutTabLocalService
